requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.3
django-extensions==2.2.1
django-werkzeug==1.0.0
djangorestframework==3.10.1
drf-extensions==0.5.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.1
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
yapf==0.28.0